+++ docs/changelog.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+[Unreleased] 1.1.0
 
 ### Added
 
 - Github CI Integration tests with ContainerLab
+- Normalized the result output between vendoers to inclue 'ok' key
 
 ### Changed